feat(s3): add unit tests and shorten ecs zip names
The ECS task name contains the cluster name, no need to have them both
in the S3 .zip object name.

this changes the names of the .zip files when using the agent for
Fargate. The exact format of the names is not pinned down so this is not
a semver break but be aware